+++ src/env.ts
@@ -1,4 +1,3 @@
-import { MsgType } from "lib/types";
 import type { ContractAddr, ChainGasPrice, Token, U } from "lib/types";
 import type { CelatoneConstants, CelatoneContractAddress } from "types";
 
@@ -49,17 +48,9 @@ export const FALLBACK_LCD_ENDPOINT: Record<string, string> = {
 
 export const MAX_FILE_SIZE = 800_000;
 
-export const MSG_TYPE_URL = {
-  [MsgType.STORE_CODE]: "/cosmwasm.wasm.v1.MsgStoreCode",
-  [MsgType.INSTANTIATE]: "/cosmwasm.wasm.v1.MsgInstantiateContract",
-  [MsgType.EXECUTE]: "/cosmwasm.wasm.v1.MsgExecuteContract",
-  [MsgType.MIGRATE]: "/cosmwasm.wasm.v1.MsgMigrateContract",
-};
-
 export const CELATONE_CONSTANTS: CelatoneConstants = {
   gasAdjustment: 1.6,
   maxFileSize: MAX_FILE_SIZE,
-  msgTypeUrl: MSG_TYPE_URL,
 };
 
 export const DUMMY_MNEMONIC = process.env.NEXT_PUBLIC_DUMMY_MNEMONIC;
@@ -84,7 +75,7 @@ export const getChainApiPath = (chainName: string) => {
 export const getMainnetApiPath = (chainId: string) => {
   switch (chainId) {
     case "osmo-test-4":
-    case "osmosis":
+    case "osmosis-1":
       return "osmosis-1";
     default:
       return undefined;