+++ src/lib/app-fns/tx/upload.tsx
@@ -26,7 +26,7 @@ interface UploadTxParams {
   memo?: string;
   client: SigningCosmWasmClient;
   onTxSucceed?: (codeId: number) => void;
-  onMigrate: boolean;
+  isMigrate: boolean;
 }
 
 export const uploadContractTx = ({
@@ -38,7 +38,7 @@ export const uploadContractTx = ({
   memo,
   client,
   onTxSucceed,
-  onMigrate,
+  isMigrate,
 }: UploadTxParams): Observable<TxResultRendering> => {
   return pipe(
     sendingTx(fee),
@@ -82,15 +82,15 @@ export const uploadContractTx = ({
               <span style={{ fontWeight: 700 }}>
                 ‘{codeDesc || `${wasmFileName}(${txInfo.codeId})`}’
               </span>{" "}
-              is available on your stored code. Would you like to instantiate
-              your code now?
+              is available on your stored code. Would you like to{" "}
+              {isMigrate ? "migrate" : "instantiate"} your code now?
             </>
           ),
           headerIcon: (
             <Icon as={MdCloudUpload} fontSize="24px" color="text.dark" />
           ),
         },
-        actionVariant: onMigrate ? "upload-migrate" : "upload",
+        actionVariant: isMigrate ? "upload-migrate" : "upload",
       } as TxResultRendering;
     }
   )().pipe(catchTxError());