+++ src/lib/app-provider/tx/upload.ts
@@ -3,22 +3,27 @@ import { useWallet } from "@cosmos-kit/react";
 import { useCallback } from "react";
 
 import { uploadContractTx } from "lib/app-fns/tx/upload";
+import type { Option } from "lib/types";
 
 export interface UploadStreamParams {
-  onTxSucceed?: (codeId: number) => void;
+  wasmFileName: Option<string>;
+  wasmCode: Option<Promise<ArrayBuffer>>;
   codeDesc: string;
+  estimatedFee: Option<StdFee>;
+  onTxSucceed?: (codeId: number) => void;
 }
 
-export const useUploadContractTx = (
-  wasmFileName: string | undefined,
-  wasmCode: Promise<ArrayBuffer> | undefined,
-  estimatedFee: StdFee | undefined,
-  onMigrate: boolean
-) => {
+export const useUploadContractTx = (isMigrate: boolean) => {
   const { address, getCosmWasmClient } = useWallet();
 
   return useCallback(
-    async ({ onTxSucceed, codeDesc }: UploadStreamParams) => {
+    async ({
+      wasmFileName,
+      wasmCode,
+      codeDesc,
+      estimatedFee,
+      onTxSucceed,
+    }: UploadStreamParams) => {
       const client = await getCosmWasmClient();
       if (!address || !client)
         throw new Error("Please check your wallet connection.");
@@ -32,16 +37,9 @@ export const useUploadContractTx = (
         fee: estimatedFee,
         client,
         onTxSucceed,
-        onMigrate,
+        isMigrate,
       });
     },
-    [
-      address,
-      getCosmWasmClient,
-      estimatedFee,
-      wasmCode,
-      wasmFileName,
-      onMigrate,
-    ]
+    [address, getCosmWasmClient, isMigrate]
   );
 };