+++ src/lib/components/modal/contract/SaveNewContract.tsx
@@ -95,7 +95,7 @@ export function SaveNewContract({ list, buttonProps }: SaveNewContractProps) {
 
   // TODO: Abstract query
   const { refetch } = useQuery(
-    ["query", "instantiateInfo", contractAddressState],
+    ["query", "instantiateInfo", endpoint, contractAddressState],
     async () =>
       queryInstantiateInfo(
         endpoint,