+++ src/lib/model/contract.ts
@@ -5,9 +5,8 @@ import { useCelatoneApp } from "lib/app-provider";
 import { INSTANTIATED_LIST_NAME } from "lib/data";
 import { useCodeStore, useContractStore, useLCDEndpoint } from "lib/hooks";
 import { useAssetInfos } from "lib/services/assetService";
-import type { InstantiateInfo, PublicInfo } from "lib/services/contract";
+import type { InstantiateInfo } from "lib/services/contract";
 import {
-  queryPublicInfo,
   queryContractBalances,
   queryInstantiateInfo,
 } from "lib/services/contract";
@@ -20,13 +19,19 @@ import {
   useTxsCountByContractAddress,
   useRelatedProposalsCountByContractAddress,
 } from "lib/services/contractService";
+import {
+  usePublicProjectByContractAddress,
+  usePublicProjectBySlug,
+} from "lib/services/publicProjectService";
 import type { CodeLocalInfo } from "lib/stores/code";
 import type { ContractLocalInfo, ContractListInfo } from "lib/stores/contract";
 import type {
   BalanceWithAssetInfo,
   ContractAddr,
+  Detail,
   HumanAddr,
   Option,
+  PublicInfo,
 } from "lib/types";
 import { formatSlugName } from "lib/utils";
 
@@ -35,7 +40,10 @@ export interface ContractData {
   codeInfo: Option<CodeLocalInfo>;
   contractLocalInfo: Option<ContractLocalInfo>;
   instantiateInfo: Option<InstantiateInfo>;
-  publicInfo: Option<PublicInfo>;
+  publicProject: {
+    publicInfo: Option<PublicInfo>;
+    publicDetail: Option<Detail>;
+  };
   balances: Option<BalanceWithAssetInfo[]>;
   initMsg: string;
   initTxHash: Option<string>;
@@ -87,16 +95,19 @@ export const useInstantiatedMockInfoByMe = (): ContractListInfo => {
 
 export const useContractData = (
   contractAddress: ContractAddr
-): ContractData | undefined => {
+): Option<ContractData> => {
   const { indexerGraphClient } = useCelatoneApp();
   const { currentChainRecord } = useWallet();
   const { getCodeLocalInfo } = useCodeStore();
   const { getContractLocalInfo } = useContractStore();
   const endpoint = useLCDEndpoint();
   const assetInfos = useAssetInfos();
+  const { data: publicInfo } =
+    usePublicProjectByContractAddress(contractAddress);
+  const { data: publicInfoBySlug } = usePublicProjectBySlug(publicInfo?.slug);
 
   const { data: instantiateInfo } = useQuery(
-    ["query", "instantiateInfo", contractAddress],
+    ["query", "instantiateInfo", endpoint, contractAddress],
     async () =>
       queryInstantiateInfo(endpoint, indexerGraphClient, contractAddress),
     { enabled: !!currentChainRecord }
@@ -126,17 +137,6 @@ export const useContractData = (
       return -1;
     });
 
-  const { data: publicInfo } = useQuery(
-    ["query", "publicInfo", contractAddress],
-    async () =>
-      queryPublicInfo(
-        currentChainRecord?.name,
-        currentChainRecord?.chain.chain_id,
-        contractAddress
-      ),
-    { enabled: !!currentChainRecord }
-  );
-
   const codeInfo = instantiateInfo
     ? getCodeLocalInfo(Number(instantiateInfo.codeId))
     : undefined;
@@ -155,7 +155,10 @@ export const useContractData = (
     codeInfo,
     contractLocalInfo,
     instantiateInfo,
-    publicInfo,
+    publicProject: {
+      publicInfo,
+      publicDetail: publicInfoBySlug?.details,
+    },
     balances: contractBalancesWithAssetInfos,
     initMsg: instantiateDetail.initMsg,
     initTxHash: instantiateDetail.initTxHash,