+++ src/lib/pages/migrate/index.tsx
@@ -21,7 +21,7 @@ import { UploadNewCode } from "./components/UploadNewCode";
 import type { MigratePageState } from "./types";
 
 const defaultValues: MigratePageState = {
-  migrateStep: 1.1,
+  migrateStep: "migrate_options",
   contractAddress: "" as ContractAddr,
   admin: undefined,
   codeId: "",
@@ -39,8 +39,13 @@ const Migrate = () => {
   });
   const { migrateStep, contractAddress, admin, codeId } = watch();
 
-  const firstStep = migrateStep !== 2;
-  const handleBack = () => setValue("migrateStep", 1.1);
+  const firstStep = migrateStep !== "migrate_contract";
+  const handleBack = () => setValue("migrateStep", "migrate_options");
+
+  const contractAddressParam = getFirstQueryParam(
+    router.query.contract
+  ) as ContractAddr;
+  const codeIdParam = getFirstQueryParam(router.query["code-id"]);
 
   const onContractSelect = useCallback(
     (contract: ContractAddr) => {
@@ -57,7 +62,7 @@ const Migrate = () => {
   );
 
   useQuery(
-    ["query", "instantiateInfo", contractAddress],
+    ["query", "instantiateInfo", endpoint, contractAddress],
     async () =>
       queryInstantiateInfo(endpoint, indexerGraphClient, contractAddress),
     {
@@ -79,19 +84,15 @@ const Migrate = () => {
   );
 
   useEffect(() => {
-    const contractAddressParam = getFirstQueryParam(
-      router.query.contract
-    ) as ContractAddr;
-    const codeIdParam = getFirstQueryParam(router.query["code-id"]);
-
     setValue("contractAddress", contractAddressParam);
     setValue("codeId", codeIdParam);
-    if (contractAddressParam && codeIdParam) setValue("migrateStep", 2);
-  }, [setValue]);
+    if (contractAddressParam && codeIdParam)
+      setValue("migrateStep", "migrate_contract");
+  }, [codeIdParam, contractAddressParam, setValue]);
 
   const renderBody = () => {
     switch (migrateStep) {
-      case 2:
+      case "migrate_contract":
         return (
           <MigrateContract
             contractAddress={contractAddress}
@@ -99,18 +100,18 @@ const Migrate = () => {
             handleBack={handleBack}
           />
         );
-      case 1.2:
+      case "upload_new_code":
         return <UploadNewCode handleBack={handleBack} />;
-      case 1.1:
+      case "migrate_options":
       default:
         return (
           <MigrateOptions
             isAdmin={admin === address}
             uploadHandler={() => {
-              setValue("migrateStep", 1.2);
+              setValue("migrateStep", "upload_new_code");
             }}
             existHandler={() => {
-              setValue("migrateStep", 2);
+              setValue("migrateStep", "migrate_contract");
             }}
           />
         );